requirements.txt
requests
beautifulsoup4
python-telegram-bot==20.6
aiohttp